heap-allocate network snapshots rather than making copies
This seems to fix the crashes that were happening when pushing a new
snapshot to the model-stats structure. Somewhere when combined with
optimizations and std::async implicitly relying on copy-constructiors
was creating a problem. Instead of passing around references and
letting default copies be created when snapshots are added to the
history we now use shared pointers and allocate snapshots on the heap.